requirements.txt
Padding==0.5
pysnappy==0.6.1
requests==2.23.0
Flask==1.1.2
pycryptodome==3.9.7
